move qlayer utils
